sway: add simple automatic layout mode
Create a very simple version of XMonad's (and dwm's) Tall layout: the first
window goes into a master pane, which currently takes the left half of the
surrounding container, filling it from top to bottom. The remaining windows
are arranged vertically, from second window downwards, filling the other
half of the container.