sway: rip out L_{HORIZ,VERT,TABBED} and related
Remove those variants, a bunch of conditional code for those cases, and
a number of functions.  In particular, remove the `split*` commands.

This code is highly experimental.  It needs to be replaced or cleaned up
at some point.  The purpose for this is mainly figuring out how to
rewrite sway to use XMonad/dwm-style automatic layouts.